them crud phong

+++ source/J2EE/src/main/java/vn/edu/uit/entity/MPhong.java
@@ -74,7 +74,7 @@ public class MPhong implements Serializable {
 		StatusID = statusID;
 	}
 
-	public double getGiaNgay() {
+	public long getGiaNgay() {
 		return GiaNgay;
 	}
